2009-08-05 Roman - !!:/UI/Vision: Interface translation to German is added by Irina Popkova. - FIX:/Archive, /UI/WebVision: LibGD2 function gdImageStringFT error open font print is added. Drawing images without text is released.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@949 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/System/da_mem.cpp
@@ -61,7 +61,7 @@ void Mem::deInit( TMdPrm *prm )
 
 void Mem::getVal( TMdPrm *prm )
 {
-    long m_total, m_free, m_buff, m_cach, sw_total, sw_free;
+    int m_total, m_free, m_buff, m_cach, sw_total, sw_free;
     char buf[256];
 
     FILE *f = fopen("/proc/meminfo","r");
